refactor: rename GraphqlContext to ResolverContextLike
