+++ src/has_headers.rs
@@ -1,12 +1,12 @@
 use hyper::HeaderMap;
 
-use crate::lambda::{EvaluationContext, GraphqlContext};
+use crate::lambda::{EvaluationContext, ResolverContextLike};
 
 pub trait HasHeaders {
   fn headers(&self) -> &HeaderMap;
 }
 
-impl<'a, Ctx: GraphqlContext<'a>> HasHeaders for EvaluationContext<'a, Ctx> {
+impl<'a, Ctx: ResolverContextLike<'a>> HasHeaders for EvaluationContext<'a, Ctx> {
   fn headers(&self) -> &HeaderMap {
     self.headers()
   }
